+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/BlockstackSession.kt
@@ -310,7 +310,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
             v8userSession.executeVoidFunction("redirectToSignIn", v8params)
             v8params.release()
         } catch (e: Exception) {
-            errorCallback(Result(null, e.toString()))
+            errorCallback(Result(null, ResultError(ErrorCode.RedirectFailed, e.toString())))
         }
     }
 
@@ -332,7 +332,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
             v8userSession.executeVoidFunction("redirectToSignInWithAuthRequest", v8params)
             v8params.release()
         } catch (e: Exception) {
-            errorCallback(Result(null, e.toString()))
+            errorCallback(Result(null, ResultError(ErrorCode.UnknownError, e.toString())))
         }
     }
 
@@ -693,7 +693,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
             val cipherObject = JSONObject(result)
             return Result(CipherObject(cipherObject))
         } else {
-            return Result(null, "failed to encrypt")
+            return Result(null, ResultError(ErrorCode.UnknownError,"failed to encrypt"))
         }
     }
 
@@ -729,7 +729,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
                 return Result(Base64.decode(plainContent, Base64.DEFAULT))
             }
         } else {
-            return Result(null, "failed to decrypt")
+            return Result(null, ResultError(ErrorCode.FailedDecryptionError,"failed to decrypt"))
         }
     }
 
@@ -809,7 +809,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
 
         fun signInFailure(error: String) {
             if (blockstackSession.signInCallback != null) {
-                blockstackSession.signInCallback!!.invoke(Result(null, error))
+                blockstackSession.signInCallback!!.invoke(Result(null, ResultError.fromJS(error)))
                 blockstackSession.signInCallback = null
             }
         }
@@ -824,7 +824,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun validateProofsFailure(error: String) {
-            blockstackSession.validateProofsCallback?.invoke(Result(null, error))
+            blockstackSession.validateProofsCallback?.invoke(Result(null, ResultError.fromJS(error)))
         }
 
         fun resolveZoneFileToProfileResult(profileString: String) {
@@ -833,7 +833,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun resolveZoneFileToProfileFailure(error: String) {
-            blockstackSession.resolveZoneFileToProfileCallback?.invoke(Result(null, error))
+            blockstackSession.resolveZoneFileToProfileCallback?.invoke(Result(null, ResultError.fromJS(error)))
         }
 
         fun lookupProfileResult(username: String, userDataString: String) {
@@ -842,7 +842,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun lookupProfileFailure(username: String, error: String) {
-            blockstackSession.lookupProfileCallbacks[username]?.invoke(Result(null, error))
+            blockstackSession.lookupProfileCallbacks[username]?.invoke(Result(null, ResultError.fromJS(error)))
         }
 
         fun getFileResult(content: String?, uniqueIdentifier: String, isBinary: Boolean) {
@@ -856,7 +856,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun getFileFailure(error: String, uniqueIdentifier: String) {
-            blockstackSession.getFileCallbacks[uniqueIdentifier]?.invoke(Result(null, error))
+            blockstackSession.getFileCallbacks[uniqueIdentifier]?.invoke(Result(null, ResultError.fromJS(error)))
             blockstackSession.getFileCallbacks.remove(uniqueIdentifier)
         }
 
@@ -866,7 +866,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun putFileFailure(error: String, uniqueIdentifier: String) {
-            blockstackSession.putFileCallbacks[uniqueIdentifier]?.invoke(Result(null, error))
+            blockstackSession.putFileCallbacks[uniqueIdentifier]?.invoke(Result(null, ResultError.fromJS(error)))
             blockstackSession.putFileCallbacks.remove(uniqueIdentifier)
         }
 
@@ -876,7 +876,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun deleteFileFailure(error: String, uniqueIdentifier: String) {
-            blockstackSession.deleteFileCallbacks[uniqueIdentifier]?.invoke(Result(null, error))
+            blockstackSession.deleteFileCallbacks[uniqueIdentifier]?.invoke(Result(null, ResultError.fromJS(error)))
             blockstackSession.deleteFileCallbacks.remove(uniqueIdentifier)
         }
 
@@ -886,7 +886,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun getFileUrlFailure(error: String, uniqueIdentifier: String) {
-            blockstackSession.getFileUrlCallbacks[uniqueIdentifier]?.invoke(Result(null, error))
+            blockstackSession.getFileUrlCallbacks[uniqueIdentifier]?.invoke(Result(null, ResultError.fromJS(error)))
             blockstackSession.getFileUrlCallbacks.remove(uniqueIdentifier)
         }
 
@@ -895,7 +895,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun getAppBucketUrlFailure(error: String) {
-            blockstackSession.getAppBucketUrlCallback?.invoke(Result(null, error))
+            blockstackSession.getAppBucketUrlCallback?.invoke(Result(null, ResultError.fromJS(error)))
         }
 
         fun getUserAppFileUrlResult(url: String) {
@@ -903,7 +903,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun getUserAppFileUrlFailure(error: String) {
-            blockstackSession.getAppBucketUrlCallback?.invoke(Result(null, error))
+            blockstackSession.getAppBucketUrlCallback?.invoke(Result(null, ResultError.fromJS(error)))
         }
 
         fun listFilesResult(url: String) {
@@ -915,7 +915,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun listFilesFailure(error: String) {
-            blockstackSession.listFilesCallback?.invoke(Result(null, error))
+            blockstackSession.listFilesCallback?.invoke(Result(null, ResultError.fromJS(error)))
         }
 
         fun listFilesCountResult(count: Int) {
@@ -923,7 +923,7 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
         }
 
         fun listFilesCountFailure(error: String) {
-            blockstackSession.listFilesCountCallback?.invoke(Result(null, error))
+            blockstackSession.listFilesCountCallback?.invoke(Result(null, ResultError.fromJS(error)))
         }
 
         fun getSessionData(): String {
@@ -1048,25 +1048,25 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
 
         const val TAG = "BlockstackSession"
 
-        fun verifyAuthResponse(authResponse: String): String? {
+        fun verifyAuthResponse(authResponse: String): ResultError? {
             try {
                 val tokenParts = authResponse.split('.')
                 if (tokenParts.size != 3) {
-                    return "The authResponse parameter is an invalid base64 encoded token\n2 dots requires\nAuth response: $authResponse"
+                    return ResultError(ErrorCode.UnknownError,"The authResponse parameter is an invalid base64 encoded token\n2 dots requires\nAuth response: $authResponse")
                 }
                 val decodedToken = Base64.decode(tokenParts[0], Base64.DEFAULT)
                 val stringToken = decodedToken.toString(Charsets.UTF_8)
                 val jsonToken = JSONObject(stringToken)
                 if (jsonToken.getString("typ") != "JWT") {
-                    return "The authResponse parameter is an invalid base64 encoded token\nHeader not of type JWT:${jsonToken.getString("typ")}\n Auth response: $authResponse"
+                    return ResultError(ErrorCode.UnknownError, "The authResponse parameter is an invalid base64 encoded token\nHeader not of type JWT:${jsonToken.getString("typ")}\n Auth response: $authResponse")
                 }
             } catch (e: IllegalArgumentException) {
-                val error = "The authResponse parameter is an invalid base64 encoded token\n${e.message}\nAuth response: $authResponse"
-                Log.w(TAG, IllegalArgumentException(error, e))
+                val error = ResultError(ErrorCode.UnknownError,"The authResponse parameter is an invalid base64 encoded token\n${e.message}\nAuth response: $authResponse")
+                Log.w(TAG, error.toString(), e)
                 return error
             } catch (e: JSONException) {
-                val error = "The authResponse parameter is an invalid json token\n${e.message}\nAuth response: $authResponse"
-                Log.w(TAG, IllegalArgumentException(error, e))
+                val error =  ResultError(ErrorCode.UnknownError,"The authResponse parameter is an invalid json token\n${e.message}\nAuth response: $authResponse")
+                Log.w(TAG, error.toString(), e)
                 return error
             }
             return null